tracker-miner-fs: Use peer-to-peer D-Bus communication with tracker-extract-3
Move towards peer-to-peer communication with tracker-extract-3, dropping this
service from the session bus. To achieve this, the TrackerExtractWatchdog object
on the tracker-miner-fs-3 side is now further in charge of tracker-extract-3,
also being responsible now for spawning it, and shutting it down. The subprocess
does not inherit file descriptors, except standard I/O and those explicitly passed.

The communication is established through a socketpair() created on the
tracker-miner-fs-3 side, a "server" D-Bus connection is set up on this side,
and the other socket FD is passed to the tracker-extract-3 process being
spawned through a new --socket-fd argument. The miner will expose a distinct
TrackerEndpointDBus and the private org.freedesktop.Tracker3.Files interface
over this connection.

The tracker-extract-3 process picks this FD, and creates a "client" D-Bus
connection, all communication with the SPARQL endpoint and other miner D-Bus
interfaces will be done through this connection.  In order to ease testing
of this process, it is still possible to launch without a --socket-fd argument,
in which case it will use the session bus and the public tracker-miner-fs-3
SPARQL endpoint.

Since the tracker-extract-3 process is no longer meant to be D-Bus autostarted,
or systemd managed, drop these service files. This is now fully a tracker-miner-fs-3
subprocess, thus inheriting everything necessary from it.

And since opening the client D-Bus connection from an already created socket
connection requires getsockopt(), allow this syscall through seccomp.