libtracker-miners-common: Restrict connect() syscall in seccomp
With the peer-to-peer communication set up in the previous commits,
the tracker-extract-3 process receives a socket FD for communication
and does not *need* to connect() to sockets anymore.

Since there are existing outliers that want to connect() to sockets
(mainly GStreamer indirectly, through its modules and dependencies),
we cannot let that fail with SIGSYS signal, so let these calls fail
with errno EACCES. This does not seem to result in failures or errors
from the observed modules.