libtracker-miners-common: Restrict socket access further
Make all socket() calls fail altogether with EPERM, we can now
do that with the peer to peer connection, and fixes the socket()
syscall saga in a way that libseccomp will know what to do on all
architectures.

Also drop the rule for shutdown() as it seems unnecessary and unwanted.

The socketpair() call still needs the existing allow rule for
AF_UNIX since this seems to be used for inter-thread communication
in GStreamer. This is however uncontroversial, since these sockets
do not inherently connect anywhere outside the process. Since
AF_LOCAL and AF_UNIX are the same number underneath, we can however
drop one of the two rules.