ci: Look for matching ref in user remote for branch pipelines
We have not enabled merge request pipelines yet on the tracker-miners
repository, so the paths looking up for CI_MERGE_REQUEST_TARGET_BRANCH_NAME
stand somewhat useless.

In addition to looking for a matching ref in the official tracker git
repository, also look up a possible matching tracker repository in the
user namespace.