tracker-miner-fs: Add TrackerFilesInterface API to rescue/restore FD
The persistence FD will keep the last file path that tracker-extract-3 was
dealing with, and is used in the case of e.g. crashes to let it skip that
file.

However, we dispose and re-create the TrackerFilesInterface each time
tracker-extract-3 is spawned. Add some API to retrieve the persistence
FD, and a constructor that provides the FD from prior runs.