tracker-extract: Reduce default task deadline duration
We currently allow 30 seconds for the extractor to handle
any given file. Reduce that to a default of 5 seconds which
still feels a generous allotment in cpu cycles for most usual
cases.

But add an environment variable check, so this can be overridden
in the test suite. This allows preserving the 30 seconds timeout
there, to cater for possibly busy CI runners.