tests: Add "misbehavior" functional test
This functional test will poke at the limits of tracker-extract-3,
and replicate different kinds of misbehavior that might ever happen
in it.

The functional test relies on its own set of directories for the
extractor rules, each of the extractor modules replicating the
misbehavior, the extractor will be made to look at these paths for
each test.

These tests are somewhat twisted, as extractor failures are expected,
and things unexpectedly going wrong are seen as blatant success. To
handle this, we let these extractor modules create a "fail://" resource
on unexpected success, and check that it does not exist later in the
test.

This also helps testing the tracker-miner-fs-3 codepaths handling
failures and tracker-extract-3 restarts, which was untested previously.

The modules added test the following situations:
- Unexpected exit (also e.g. crashes)
- Stalled extraction
- Modules succeeding but returning no data
- Modules not succeeding, but not providing an error
- Modules inserting wrong RDF data

These tests do run with a reduced extractor deadline of 5 seconds,
in order to avoid CI timeouts when testing the 'stalled' test.