build: Check for landlock being enabled in kernel at build time
Besides checking that the kernel landlock headers are installed, also
check that the feature is enabled in the kernel at build time.

Since landlock has not been made mandatory at the moment, build has only
been made to error out if the landlock auto build option was forced on,
otherwise it will result in the feature being implicitly disabled, and
some warning words at the end of setup: "Distribution is discouraged".

Closes: https://gitlab.gnome.org/GNOME/tracker-miners/-/issues/298