tracker-miner-fs: Wait for overall idle state before releasing heap
Currently, we attempt to release heap after the TrackerMinerFiles
went through all the files. However this is also the time when we
possibly spawn the extractor and let it go through unextracted files.
Of course, releasing the heap meanwhile the extractor is operating
(i.e. causing batch updates on the tracker-miner-fs-3 side) is
from pointless (heap space will be needed again) to "bad" (we have
to ask for this heap memory again).

But nowadays, the miner knows of the extraction progress, and forwards
it through TrackerMiner::status as its own. We can track these status
changes, and defer this timeout to a point where the whole miner
(incl. extractor) are idle.

This makes the heap release operation after miner activity happen
at a more effective and less awkward time.