libtracker-miners-common: Ensure that SIGSYS gets propagated
Issue tgkill(..., SIGSYS) in our SIGSYS signal handler to ensure
that the signal will be propagated. We nowadays indirectly get
that thanks to sigreturn being a disallowed syscall, which results
in signal recursion and terminates the process with a core file.

Generating a core file with a backtrace pointing to the failed
syscall is desirable, so do this directly with less of a roundabout.