libtracker-miners-common: Allow reading tracker-extract-3 executable
We need EXEC+READ permissions on the extractor executable, or the
"tracker3 extract" CLI tool will fail trying to execute it with
EACCES errors.

It is a bit of a mystery why we require this from the CLI tool,
but tracker-miner-fs-3 can launch its private instance without issues.
But let's roll with it.

Closes: https://gitlab.gnome.org/GNOME/tracker-miners/-/issues/310