tracker-miner-fs: Synchronize startup of private bus connection
Creating the connection and setting up the endpoint on the miner side
asynchronously may result in errors like:

  (tracker-extract-3:810568): Tracker-CRITICAL **: 14:06:28.604: Could not connect to filesystem miner endpoint: Object does not exist at path “/org/freedesktop/Tracker3/Endpoint”
  (tracker-miner-fs-3:800395): Tracker-WARNING **: 15:06:28.606: Extractor subprocess died unexpectedly: Child process exited with code 1

Since the process spawned gets ahead at trying to poke the endpoint before it is
created on the tracker miner side. This is almost harmless since the extractor
process is restarted and presumably succeeds to connect at some point.

Still not nice to have, so block message processing on the bus until
the `tracker-miner-fs-3` side raised the necessary interfaces and object
paths. This will make the extractor process "block" until everything
is set up.