tracker-writeback: Do not include specific libtracker-sparql files
It is an accident both that the SPARQL library would allow including the
libtracker-sparql/tracker-ontologies.h header specifically without going
through the main tracker-sparql.h header, and that it is used like that
here.

This file is already a dependency of the main tracker-sparql.h header,
so drop this direct include and rely that the ontology prefixes will be
pulled through there.