cli: Unite all commands into a single executable
This CLI tool does not need extensibility, so drop the scheme
inherited from Tracker/TinySPARQL where subcommands where their
own executable described by a desktop file, and unite them into
a single binary.

The only exception is the "localsearch test-sandbox" subcommand
which is implemented as a python script installed in a special
location.