extractor: Add mechanism to get content identifier from the indexer
The mechanism to get a stable content identifier is going to change to
require udev in later commits, it makes sense to shuffle its generation
so it is solely responsibility of the indexer process, while the extractor
does not need to figure these out on its own.

To do this, let the indexer insert the base nie:InformationElement
resource in the data graph, the metadata extractor can use the IRI of
this resource and extend it, or add additional InformationElement resources
for multi-part content. This way, the extractor can just query a proper
IRI, instead of having to poke at filesystems to get the right information.