indexer: Use udev to generate content identifiers
Unfortunately, it seems the blkid library is not completely helpful
without root privileges, since there are no guarantees that the cache
file will be generated.

Use udev for this task, as these UUIDs will be populated as udev data,
also prefer the "UUID_SUB" property if it exists, in order to get correct
unique identifiers for e.g. btrfs subvolumes.

This piece of infrastructure has been also moved to
tracker-miner-files-methods.c, so that code changes to it are also
detected, and trigger a re-index of the affected content.

Closes: https://gitlab.gnome.org/GNOME/localsearch/-/issues/345
Closes: https://gitlab.gnome.org/GNOME/localsearch/-/issues/357