utils: Isolate graph using FROM as per spec
Using FROM NAMED did not really isolate this graph for it to
be matched via a triple pattern with no GRAPH{} clause. But
since the spec-correctness fixes in tinysparql 3.8.alpha using
FROM NAMED with no FROM clauses will also clear the anonymous
graph set, resulting in the triple pattern matching nothing.

These queries do want to use FROM instead, to offer this graph
alone through the anonymous graph set, so it can be matched by
a triple pattern not under a GRAPH{} clause.