ci: Build/install TinySPARQL separately for Coverity
Avoid going through the subproject route here, and partly restore
the ability to check out and install TinySPARQL independently that
was removed in commit 07d63010e. This is necessary for Coverity
not to analyze the TinySPARQL code here in addition.

We do not need the ability to pick a special TinySPARQL branch, so
just pick the main branch.

Fixes: 07d63010e ("ci: Drop glue to checkout a tinysparql repo with matching branches")