common: Make SIGSYS handle explicitly one-shot
We intend to take down the process with the same SIGSYS signal
result, hint the SIGSYS handler with SA_RESETHAND that will reset
the signal handler to the default after first run, making it possible
for us to raise SIGSYS again with a meaningful backtrace.

Without this, some operations caught in the sandbox like the previous
commit's would not result in a coredumpctl backtrace. Backtraces
are always useful here.