indexer: Prepend directories to be crawled
When recursively crawling a folder, prepend the folders to be
crawled. This is a seemingly small change with some consequences,
the most evident one is that folders will now be indexed in
post-order, as opposed to the level-order achieved in practice with
the previous approach.

Since most relevant data is usually close to the root folders in
filesystems this is arguably a less nice behavior. However there
are 2 less evident changes that this brings and are desirable:

- Previously, if there was be a folder with a large amount of
  child directories, these folders found would queue up to be
  processed for a maximum amount of time. With the new behavior
  these items will get processed and flushed ASAP leading to less
  memory usage overall in these situations.
- There will be better ordering guarantees if we wanted to run something
  when a folder (and its contents) are fully guaranteed to be up-to-date.
  This is an advantage that will be used in later commits.