indexer: Complete folders as a last step
Until now, we used to process parent folders upfront, this brought
the opportunity for us to end up missing files to be processed if
the process got interrupted in the right moment (e.g. shutdown or
crash). This situation would get eventually corrected (e.g. through
monitor events), but not immediately.

To fix this, use the underlying TrackerFileNotifier::finished-folder
signal plumbing in order to "finalize" parent folders as a last step.
This will make these folders be still seen as crawl-worthy on the
next process restart, ensuring that no content is missed.

This improves robustness in the case of crashes, etc, and will work
however we split the indexed data in batches/transactions.