extractor: Proxy progress through extractor D-Bus interface
The code cleanups at commit 885b30d0628 went a bit overeager since
the extractor process still used the miner's D-Bus interface
Progress signal through TrackerMinerProxy in order to communicate
progress to the main indexer process.

This is no more, so proxy the extractor progress through a signal
in our private interface for communication with the indexer, and
make the indexer's TrackerExtractWatchdog look for the signal in
the new location.

This makes progress from the extractor forwarded again, e.g. to
"localsearch daemon -f".

Fixes: 885b30d06 ("extractor: Code cleanup")