extractor: Plug leaks on JPEG extractor failures
As longjmp-reliant libraries go, the stack will be "restored" to the
original state on failure, meaning that any change done to it will
be forgotten, and the references in stack to the memory allocated
between the setjmp and the failure possibly lost.

Shuffle the TrackerResource creation, so that it is not created at
that failable time, and perform its creation after all failable
libjpeg operations are done. This avoids the TrackerResource leak
on libjpeg failures.