indexer: Simplify queue processing status
Given we process elements incrementally, the situation where the
TrackerFileNotifier is done, but there are items left to be processed
are basically reserved to the very end of the processing queue.

This is the first time ever where we have a full picture of the items
that are left/indexed, this situation is however so narrow that it does
not make sense to report it, so also drop the calculation of
elapsed/remaining times.