tests: Add functional test for offline folder updates
Test that folders are completely updated (i.e. their nie:InformationElement gets
deleted and replaced) if a folder is updated while the miner is offline. It did
not use to be the case, resulting in stale nfo:Folder resources being left in the
database.

This did not result in warnings by itself, but could result in errors like:

Could not execute sparql: Unable to insert multiple values on single valued property `nfo:belongsToContainer' for resource file:///... (old_value: '593', new value: '603')

When a file on an affected folder got moved.