indexer: Avoid memory payload in handling folder updates
Prior to this commit we used to handle updated folders found during
crawling as:

- While iterating cursor results (i.e. known existing elements) we'd
  detect the folder update, and queue it for crawling.
- Changes found in child files of an updated folder are for the most
  part accumulated until the folder is crawled and not processed
  directly, even though we stat() those files in place and are able
  to detect file updates, deletions, and non-updates. We just cannot
  detect new files at this point.
- After the folder is crawled and all new elements are found, the
  whole set of changes to child files is handled.

This can create scenarios where we accumulate a large amount of data
for these child files in memory until crawling. Pathologically bad
cases would be handling updates on a very large folder, or the user
doing `find . -type d -exec touch -m {} \;` to change mtime for
every folder, and this to be handled on next indexer restart.

To change this, avoid having to preserve file information between these
two phases, by:

- Making the cursor iteration phase handle in place all changes to
  known files, including file updates and deletions. This is consistent
  with the handling of file changes that did not happen in a folder that
  was updated itself.
- Exclusively delegating finding new files on the crawling phase. This
  is done by a per-file ASK query to bail out on already known content.

This avoids having to keep indefinitely large amount of data between those
two phases in memory, keeping memory usage flatter. And it is even
slightly faster on pathological cases with many updated folders, since we
replace a roughly O(n^2) traversal with a database lookup through an index.