extract: Switch to ffmpeg/libav for audio/video metadata extraction
Now that the libav-based extractor is quite on par with the GStreamer
one, switch to using it by default for all audio/video formats, and
drop the GStreamer-based extractor.

The reasons to do this are various:
- GStreamer's own modular support runs at odds with our extractor
  sandbox. Its attempts to recreate the registry file cannot be
  fully stopped for people doing live system updates.
- GStreamer's own modular support also means it's a whack-a-mole
  game to inhibit all those undesirable modules (e.g. video acceleration,
  webcams, ...) unrelated to metadata extraction.
- GStreamer is designed for the long run, e.g. opening a file and operating
  on it. Turning metadata extraction into a race across the filesystem
  is very suboptimal for it. The libav extractor is around 2.5x faster
  for the same set of files here locally (~73s for 8240 files, compared to
  ~175s with GStreamer)
- Possibly for the same reasons as the previous point, the memory usage
  as seen by top for the localsearch-extract-3 is significantly lower with
  the libav extractor.

It turns out that GStreamer would resort to its own libav extractor
for many formats, and even in the cases it did not FFmpeg seems quite
capable of handling them. There does not seem to be any major loss of
functionality.

Some tests had to update their expected output, notably:
- test_extractor_flac_cuesheet now has the right nfo:duration for the
  first track.
- mkv_basic lacks nfo:duration, possibly a ffmpeg bug since other
  formats than MKV work.