indexer: Check high water more often after dequeue
We check once after flush, which might not happen timely. Instead
checking after each item being dequeued for processing, to keep it
symmetric with the check happening when we queue up elements.

Since we've somewhat decoupled the number of files in the queue from
the number of operations in the SPARQL queue (there may be several
per file), we have otherwise some room for the file notifier to be
stuck in high water and leave the higher layers waiting for more.