indexer: Do not early out if getting content type failed
In the same TOCTOU situations as last commit we might end up with
a directory that was deleted and got no related nfo:FileDataObject
information written to the database, which would result in the
following warning when the directory is signaled as finished:

(localsearch-3:1234): Tracker-WARNING **: 19:06:48.805: Could not execute sparql: Subject file:///... is not is not a nfo:FileDataObject, cannot have property `tracker:extractorHash'

To fix this, avoid ignoring early any process_file*() request on
already deleted files, and rely on the deletion we should typically
obtain from filesystem monitors immediately afterwards. The folder
data will be still batched into the database meanwhile, so the
::finish-directory() operation will find the expected resource with
the expected rdf:type.