localsearch: Drop case for running on session bus
This was left as a debugging method, but nowadays the recommended
way to instrument LocalSearch is by running the indexer and tracing
children as well, as we have the tools to do so (e.g. valgrind checks).

Drop this way to run localsearch-extract-3 from the commandline as
a session bus client, and mandate that the socket FD is used.