cli: Handle URIs not under namespaces in "localsearch info"
The namespace manager returns NULL if it tries to compress an
URI that is not under the existing namespaces, so this will evidently
fail for misc string/iri references to HTTP URLs.

Fall back, and use the full string in the output in that case, in
order to avoid crashes handling this kind of data.