indexer: Improve/fix TrackerLRU management in item_move()
With the item_move() logic completely laid out within the same function,
it is clearer that the LRU might be left with stale file<->identifier
relations for elements in the source/dest locations in some conditions.

Make the LRU management more self-contained there, and more obviously
correct at a glance. This also meant we can tighten the ::remove_children()
vfunc call to only apply in the necessary conditions.

Closes: https://gitlab.gnome.org/GNOME/localsearch/-/issues/392