cli: Make fts:snippet part of --detailed output in "localsearch search"
Providing the snippet for possibly thousands of results is a potentially
slow task. Make the search subcommand resolve only the "what" by default
so the fts:match queries are as fast as possible, and have the more
expensive snippet info be provided only if --detailed is passed.

This does also make the inverted --disable-snippet option pointless, so
remove it.