cli: Simplify "localsearch search" output
Avoid pointless use of colors for printing results, and the weird
padding at the beginning of each row. Since we don't print colors
anymore (except in snippets under --detailed), we can also remove
the --disable-color option.

We now just use ANSI commands to set bold text on the header, and
simply avoid it if the command is e.g. being piped to a file.