extractor: Do not try to propagate a NULL error on empty text files
This error check was changed at 3bef7fb0bf96 in order to appease
Coverity, but there are situations where a NULL content can be gotten
with no error associated, e.g. files with an indeterminate encoding.

Revert the check, and rely on g_auto to make everything more reassuring
to Coverity.

Fixes: 3bef7fb0b ("extractor: Make error check more appealing to Coverity")