indexer: Forward enable-monitors setting directly to TrackerMonitor
Let the setting control the TrackerMonitor::enabled property for the
global monitor, instead of passing it as TRACKER_DIRECTORY_FLAG_MONITOR
individually for each configured folder.

This allows for less churn on config changes, since we don't have to
update the TrackerIndexingTree configured folders just for the sake
of setting the flag.