indexer: Deprecate crawling-interval setting
Nowadays, drive speeds and LocalSearch are both fast enough that
it does not make sense to try to avoid mtime checks in a span of
days. On the contrary, this can only lead us to showing vastly
outdated search results.

Drop this setting, and always check whether there is new content to
be indexed on startup.