indexer: Minor refactor
Long long ago, we had a configuration UI with delayed apply, and
it made some sense to coalesce filter updates. Nowadays, these settings
will be changed individually with no chance that this idle function will
be able to compress user changes. Just go with it, and handle the filter
changes individually.