indexer: Deprecate "low-disk-space-limit" setting
There's multiple reasons why this is dubious:

- Nowadays, the disparity between usual drive sizes and regular
  LocalSearch database sizes is large enough that it's a well
  intentioned but kind of pointless way to shoot ourselves in
  the foot.
- This check is insufficient, SQLite also barfs if /tmp is full
  which we don't observe here.
- Even if the disk is full, the database might still have free
  pages allocated that make it possible to continue further.

Replace this setting with a simpler check, if the NO_SPACE database
error is obtained, LocalSearch will pause itself, and resume itself
after 15 minutes, with the premise that it will stay in this long
wound loop if the disk remains full (i.e. next insertion fails again).