indexer: Deprecate "initial-sleep" property
Nowadays, the default timeout of 15 seconds is quite on the long
side, given session startup times. Also the maximum 16-17min wait
allowed by the setting is fairly ludicrous, it always was.

Replace this configurable timeout with a fancier mechanism that:
- If the service is running on a TTY (say, testing) it starts
  immediately
- If a session systemd1 manager is available it will be queried
  for session status, and only start when the session was already
  finished loading.
- If it all fails, the service will start within a hardcoded
  timeout of 5 seconds.

This will optimistically make LocalSearch wait strictly the time
necessary to start on most situations.