indexer: Deprecate "removable-days-threshold" setting
This should be eventually completely refactored away when we move
forward with the plan about per-volume databases. Meanwhile,
deprecate the option, and keep it as a fixed timeout.

This also allows us to drop the last checks on "index-removable-devices",
since these just existed to either preserve or delete indexed content
immediately.

A functional test needed changes, since it was relying on this setting
to fully remove a directory after "unmount". The test was changed to just
look for tracker:available instead.