indexer: Lazily populate root IDs in TrackerIndexingTree
Obtain lazily the relevant filesystem ID from a root. As this is
the basis for our stable identifiers, and is code that requires a
reindex if changed, it's been separated to its own *-methods.c file
so it can be handled together with the other sensitive miner methods
wrt extractor hash changes.

This getter does also expand paths (e.g. resolve symlinks) in order
to look for the correct filesystem identifiers, e.g. in the case of
Silverblue and $HOME being /home/user but this folder being really
a symlink to /var/home/user.