extractor: Fix possible crash during error recovery
If we are handling a file that failed extraction previously, we
will attempt to handle the changes first using the extractor's
internal TrackerBatch, but not logging anything on the commit
buffer.

If an error happens as a result of that batch execution, the
error handling paths will try to re-process the commit buffer one
by one synchronously, but fail as there is no commit buffer associated
to this batch.

We should only do that fallback error handling for batch executions
that result from normal file processing, and during handling of a
previously failed file we can do nothing else than moving forward and
hoping for the best, the batch does contain updates for this single
individual element.

While at it, check out on external conditions, like missing disk
space, or the database being corrupt.

Related: https://bugzilla.redhat.com/show_bug.cgi?id=2375692