indexer: Honor ignore-* setting changes with indexed files
Both when handling a file/directory filter configuration change during
the lifetime of the process and after a restart, the paths handling
whether already indexed files are up-to-date would not check whether
the files should still be eligible for indexing per the configuration
filters.

Honor this, so that files being suddenly filtered out due to
configuration changes get effectively removed from the database.