indexer: Drop size check before inserting file
Nowadays there is the application/x-zerosize content type for
empty files instead of text/plain, so there is no danger of
inserting unintended files.

This also introduced some race conditions, since we may get
an initial GFileInfo with 0 size even though the file does
have content in it already at the time of processing.

This is expected to improve robustness of the test suite, since
most timeouts in CI are due to this situation.