indexer: Fix signal emission arguments
When renaming a file, we would in some circumstances handle it
as a file update on the destination of the move operation. In this
situation we emitted a ::file-updated signal with a missing boolean
argument for whether it was an attributes-only update. This would
result in unpredictable behavior in the handler.

Fix this signal emission, so it specifies all the intended arguments.
This fixes heisenbugs, in CI and presumably in user setups.