indexer: Clear extractor process state after termination
This fixes a race condition where we might have signaled the
current extractor process to terminate, and handle changes that
make it necessary an extractor running again.

In that case tracker_extract_watchdog_ensure_started() might
still see we have a GSubprocess, all while the extractor process
is still handling SIGTERM. In this case, we'll now leave the old
instance to terminate itself away while we launch a second
instance.