extractor: Ensure to schedule a query restart on changes
Sometimes in the test suite, we startup the extractor, it queries
for data as part of the startup process, then it receives notifier
updates on changes. The initial query happened at a time that the
results come back empty, and the notifier updates try to trigger
a query restart but we are already querying.

We cannot rely that the ongoing query will catch the elements we
got notified upon, so add the mechanism to schedule a second query
to be completed after the ongoing one finishes.