extractor: Drop extra elapsed time checks for PDF text extraction
Drop the in-extractor elapsed time checks for plain text extraction,
and rely solely on the watchdog that is looking for stalled extraction.

This will make PDF plain text extraction all-or-nothing and the
extracted metadata always consistent (as long as it's extracted),
which is arguably a more consistent behavior than returning an
indeterminate amount of text.

Related: https://gitlab.gnome.org/GNOME/localsearch/-/merge_requests/541