indexer: Simplify the implementation of TrackerStorage
This object does nowadays a lot of things that we no longer need it
to do, such as optical storage tracking, and a convoluted tree structure
to carry accounting of mounts.

We do not need much of that nowadays, we can rely on GDrive to tell us
whether the drive is removable or not, and we can simplify many of its
internals so that it's more to the point.