extractor: Use partial reads for WebP image data
All the data we look for is in the image header. While there does
not seem to be a fixed/maximum size for it, reading the first 256KB
should be plenty to get the data we are looking for.

While at it, port to using the less error-prone GInputStream and
GBytes to read the data, and use g_autoptr some more.