extractor: Port RAW extractor to gexiv2 0.16, with backwards compat paths
Gexiv2 0.16 contains incompatible API changes, and even though it is
parallel installable with < 0.16, we do not want to hold its adoption.

So this commit changes the RAW extractor code to use the API as defined
in gexiv2 0.16, and makes meson prefer gexiv2 0.16 over older versions.
Since we do not mandate any specific gexiv2 version, the RAW extractor
was also added #define glue for backwards compatibility with older
gexiv2 versions, for the pre-0.12.2 metadata getters with no errors
(that we used so far), and for the "try" variants with GError arguments
that deprecated this API between 0.12.2 and 0.16.0. This makes us
support all gexiv2 API flavors.

Regarding the new GError arguments, these are of no use for us since
we want to try all metadata there is to be extracted, we would neither
bail out nor log/propagate the error, so NULL is passed everywhere
and missing values are handled as they should.

Closes: https://gitlab.gnome.org/GNOME/localsearch/-/issues/238